Remove unnecesssary state tracker locking
Now that each API call locks the calling thread and the state writer
locks all threads, state tracker does not need to lock when updating
state for API calls and their objects that require external
synchronization. Add a mutex and lock for VulkanStateWriter's internal
state table which may be modified in parallel.

Change-Id: I18e3c01b5673f439e8b848e49c87c8b963b59192