Fix D3D12 debug layer warning about resource state
Dx12ResourceDataUtil::ExecuteCopyCommandList sets up a command list to
copy data to or from a resource. In the command list are the state
transitions required to do the copy and then restore the resource to its
original state. When the command list is executed, the resource is
transitioned to D3D12_RESOURCE_STATE_COMMON then when
Copy[Buffer|Texture]Region is called it is automatically promoted to
D3D12_RESOURCE_STATE_COPY_[DEST|SOURCE]. After that, the final transition
would attempt to return the resource back to its original state
using D3D12_RESOURCE_STATE_COMMON as the "before" state of the transition,
but because the resource state was automatically promoted to
D3D12_RESOURCE_STATE_COPY_[DEST|SOURCE] this resulted in debug layer
errors.