dx12 dump: Duplicated command list
It's not good to add copy resources commands on the target command list.
The target command list might not be reset and used on multiple
ExecuteCommandLists. In the case, the copy resources commands could
be executed multiple times, and waste performance. Creating a duplicated
command list is better. The target ExecuteCommandLists run the duplicated
command list instead.