Add the `gfxrecon-replay --recapture` feature
Using `gfxrecon-replay` with the `--recapture` option causes replay to
recapture the current file. Recapture has the same options as standard
capture and the options are set using environment variables. For
recapture, the capture code is linked directly into `gfxrecon-replay`--
no GFXR capture layer is added to the Vulkan layer chain.